[sim_controllers] Move more implementation out of `sim.[h/c]pp`. Log point at which a controller causes a profileset to bail out, as well as reason.
